+++ modify_ongoing_task_result.go
@@ -3,7 +3,7 @@ package ravendb
 // ModifyOngoingTaskResult represents a raven server command
 // for modyfing task result
 type ModifyOngoingTaskResult struct {
-	TaskID           int    `json:"TaskId"`
-	RaftCommandIndex int    `json:"RaftCommandIndex"`
+	TaskID           int64  `json:"TaskId"`
+	RaftCommandIndex int64  `json:"RaftCommandIndex"`
 	ResponsibleNode  string `json:"ResponsibleNode"`
 }
