+++ size.go
@@ -2,6 +2,6 @@ package ravendb
 
 // Size describes size of entity on disk
 type Size struct {
-	SizeInBytes int    `json:"SizeInBytes"`
+	SizeInBytes int64  `json:"SizeInBytes"`
 	HumaneSize  string `json:"HumaneSize"`
 }
